pyproject.toml
[project]
name = "mann-kendall-automated"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Gabriel Barbosa Soares", email = "gabrielclimb@gmail.com" },
]
dependencies = [
    "streamlit>=1.38.0",
    "xlrd>=2.0.1",
    "xlsxwriter>=3.2.0",
    "pandas>=2.2.2",
    "scipy>=1.14.1",
    "openpyxl>=3.1.5",
    "plotly>=5.24.0",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "celery>=5.4.0",
    "django-cors-headers>=4.6.0",
    "psycopg2-binary>=2.9.10",
    "dj-database-url>=2.3.0",
    "django-celery-results>=2.5.1",
    "djangorestframework-simplejwt>=5.3.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.3",
    "uv>=0.4.2",
    "pytest-django>=4.9.0",
    "pytest-asyncio>=0.25.0",
    "python-dotenv>=1.0.1",
]

[tool.rye.scripts]
req = "uv pip compile pyproject.toml -o requirements.txt"
stream = "streamlit run app.py"
test = "pytest src/ --cov=src --cache-clear"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mann_kendall_automated"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = "test_*.py"
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]

[tool.coverage.run]
source = ["src"]
omit = ["*/migrations/*", "*/tests/*", "*/settings.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
